method map() and try catch move in separete function and adjust space
