+++ src/App.module.scss
@@ -1,28 +1,26 @@
 .App {
-  overflow: hidden;
- width: 1280px;
-  display: flex;
-  flex-direction: column;
-  
+    overflow: hidden;
+    width: 1280px;
+    display: flex;
+    flex-direction: column;
+
 }
 
-@media(max-width:1280px){
-  .App{
-    width: 100%;
-  }
+@media(max-width:1280px) {
+    .App {
+        width: 100%;
+    }
 }
 
 .main_container {
-margin: 10rem;
-display: flex;
+    margin: 10rem;
+    display: flex;
 
 }
 
-.posts{
-  margin-left: 10%;
- font-size: 1.5rem;
- width: 90vh;
- color: black;
+.posts {
+    margin-left: 10%;
+    font-size: 1.5rem;
+    width: 90vh;
+    color: black;
 }
-
-
