+++ src/app/pages/HomePage.jsx
@@ -2,11 +2,11 @@ import React from 'react'
 import Posts from '../../components/posts/Posts'
 
 const HomePage = () => {
-  return (
-    <div>
-        <Posts />
-    </div>
-  )
+    return (
+        <div>
+            <Posts />
+        </div>
+    )
 }
 
 export default HomePage
