+++ src/app/pages/NotFoundPage.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 const NotFoundPage = () => {
     return (
         <div>
-            <h1 style={{textAlign: 'center'}}>This page doesn't exist. Go <Link to="/">home</Link></h1>
+            <h1 style={{ textAlign: 'center' }}>This page doesn't exist. Go <Link to="/">home</Link></h1>
         </div>
     )
 }
