+++ src/app/pages/SubreditPage.jsx
@@ -4,13 +4,13 @@ import Posts from '../../components/posts/Posts'
 import SubredditsHeader from '../../components/subredditsHeader/SubredditHeader'
 
 const SubreditPage = () => {
-  const { status } = useSelector(state => state.posts)
-  return (
-    <div>
-      {status === 'resolved' && <SubredditsHeader />}
-      <Posts />
-    </div>
-  )
+    const { status } = useSelector(state => state.posts)
+    return (
+        <div>
+            {status === 'resolved' && <SubredditsHeader />}
+            <Posts />
+        </div>
+    )
 }
 
 export default SubreditPage
