+++ src/app/slicers/searchTermSlice.jsx
