+++ src/components/filter/Filter.css
@@ -10,10 +10,11 @@
     padding: 1.5rem;
     margin-bottom: 2rem;
 }
+
 .noActive:hover {
     cursor: pointer;
 }
 
-.filter_container img{
+.filter_container img {
     width: 1.5rem;
 }
