+++ src/index.js
@@ -13,12 +13,12 @@ const root = createRoot(container);
 root.render(
 
     <Router>
-      <Provider store={store}>
-      <App />
-    </Provider>
+        <Provider store={store}>
+            <App />
+        </Provider>
     </Router>
-    
- 
+
+
 );
 
 // If you want to start measuring performance in your app, pass a function
